Fix issue with vulkan download hash check
DepHash was set to DepVersion. For all other deps, this is fine. For
Vulkan, we actually check the hash of the installer file, so it breaks.
Actually specify and check for a third array element for the hash to
handle this case. If the third element is not present, continue to use
the Version as the Hash as before.