fix libpng, libvpx, and ffmpeg again
Swapping the error mode didn't help. For libpng and libvpx, we're in the
build dir, so the Makefile check is fine. FFmpeg ships with a Makefile,
so we need to ignore its errors and its request to stop. We could
probably use -i or --ignore-errors instead, but this works for now.