add function to get unique PATH
Deduplicate PATH (Env:Path) to bring the PATH on GHA down from 215 items
at 7805 characters to 73 items at 2652 characters. This is important
because the limit of PATH on Windows is 8191 characters. Running any
scripts that add more items to PATH runs the risk of hitting that limit
with GHA's PATH variable starting at such a large size. We can use this
function to create a stripped down PATH with only unique entries with
the order preserved. We can then replace the PATH entirely, and
optionally swap back to the original PATH if needed. This is less
troublesome than trying to manually remove unneeded items from PATH, but
that is also an option to further reduce PATH if needed.

See:
https://docs.microsoft.com/en-us/troubleshoot/windows-client/shell-experience/command-line-string-limitation