Constrain backup CI var searches to specific job
If environment variables with the same name exist in multiple different
jobs in the same workflow, the current script has no guarantee to target
the environment variables in the macOS jobs. While this is fine if the
jobs use the same versions and hashes for the dependencies, this causes
failures if the jobs use different versions or hashes. This issue was
discovered while preparing the Windows jobs in the same workflow file as
the macOS jobs. While the Windows and macOS jobs are no longer planned
to be in the same workflow file, this is probably a good future-proofing
fix to make now.