CI: Improve Qt caching
Use the hash of the Qt patches for the cache key for the Qt build jobs.
Use restore-keys to do a fuzzy-match to restore the cache if the exact
cache key is not found. This should allow use to restore the Qt cache
and do a partial rebuild if the patches used for Qt change.